Изменена "Занятие 2 (A), E.Фибоначчи возвращается"
Ничёсе там цифирки... посмотрим что даст эта библиотека. Хмм... Теперь превышение времени.

+++ Тренировка по аглоритмам/2.0/Занятие 3/Дивизион А (Олимпиадный)/E. Фибоначчи возвращается.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 def main():
     '''
     '''
@@ -5,7 +7,7 @@ def main():
 
     A = [0] * N
     for i in range(N):
-        A[i] = int(input())
+        A[i] = Decimal(input())
         
     max_number = max(A)
